git-svn-id: svn://svn.sphinxsearch.com/sphinx/trunk@4603 406a0c4d-033a-0410-8de8-e80135713968

+++ src/sphinxint.h
@@ -1927,7 +1927,7 @@ struct InfixBlock_t
 		const char *	m_sInfix;
 		DWORD			m_iInfixOffset;
 	};
-	int			m_iOffset;
+	DWORD				m_iOffset;
 };
 
 
@@ -1939,7 +1939,7 @@ public:
 	virtual			~ISphInfixBuilder() {}
 	virtual void	AddWord ( const BYTE * pWord, int iWordLength, int iCheckpoint, bool bHasMorphology ) = 0;
 	virtual void	SaveEntries ( CSphWriter & wrDict ) = 0;
-	virtual int		SaveEntryBlocks ( CSphWriter & wrDict ) = 0;
+	virtual int64_t	SaveEntryBlocks ( CSphWriter & wrDict ) = 0;
 	virtual int		GetBlocksWordsSize () const = 0;
 };
